tasks policy done

+++ app/Http/Controllers/TaskController.php
@@ -27,6 +27,7 @@ class TaskController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Task::class);
         $projects = $this->projectService->getProjectsForDropdown();
         $users = $this->userService->getUsersForDropdown();
         return view('tasks.create', compact('projects', 'users'));
@@ -34,18 +35,21 @@ class TaskController extends Controller
 
     public function show(Task $task)
     {
+        $this->authorize('view', $task);
         $task = $this->taskService->getTaskDetails($task);
         return view('tasks.show', compact('task'));
     }
 
     public function store(StoreTaskRequest $request)
     {
+        $this->authorize('create', Task::class);
         $this->taskService->createTask($request->validated());
         return redirect()->route('tasks.index')->with('success', 'Task created!');
     }
 
     public function edit(Task $task)
     {
+        $this->authorize('update', $task);
         $projects = $this->projectService->getProjectsForDropdown();
         $users = $this->userService->getUsersForDropdown();
         return view('tasks.edit', compact('task', 'projects', 'users'));
@@ -53,12 +57,14 @@ class TaskController extends Controller
 
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        $this->authorize('update', $task);
         $this->taskService->updateTask($task, $request->validated());
         return redirect()->route('tasks.index')->with('success', 'Task Updated Successfully.');
     }
 
     public function destroy(Task $task)
     {
+        $this->authorize('delete', $task);
         $this->taskService->deleteTask($task);
         return redirect()->route('tasks.index')
             ->with('success', 'Task deleted successfully!');
